refactor(e2e): ページテストをシナリオ化し不要なものを削除
記事閲覧の導線(一覧→詳細→戻る/実記事を開く、モバイルの概要展開)を
scenario配下へ移動。表示・レイアウト・カウント確認やフィルター適用フローは
E2Eのシナリオに該当しないため削除し、それに伴い未使用となったフィルターPOMと
TrendsPageの未使用メソッドも整理。
