+++ application/packages/e2e/src/pom/trends-page.ts
@@ -2,20 +2,12 @@ import { expect, type Locator, type Page } from '@playwright/test'
 import { TIMEOUT } from './constants'
 
 export class TrendsPage {
-  private readonly noArticlesMessage: Locator
   private readonly articleCards: Locator
-  private readonly qiitaIcons: Locator
-  private readonly zennIcons: Locator
-  private readonly hatenaIcons: Locator
 
   constructor(private readonly page: Page) {
-    this.noArticlesMessage = page.getByText('記事がありません')
     this.articleCards = page
       .getByRole('button')
       .filter({ has: page.getByRole('img', { name: /(?:qiita|zenn|hatena) icon/ }) })
-    this.qiitaIcons = page.getByRole('img', { name: 'qiita icon' })
-    this.zennIcons = page.getByRole('img', { name: 'zenn icon' })
-    this.hatenaIcons = page.getByRole('img', { name: 'hatena icon' })
   }
 
   async goto(path = '/trends'): Promise<void> {
@@ -31,10 +23,6 @@ export class TrendsPage {
     await expect(this.firstArticleCard()).toBeVisible({ timeout: TIMEOUT })
   }
 
-  async expectNoArticlesMessage(): Promise<void> {
-    await expect(this.noArticlesMessage).toBeVisible({ timeout: TIMEOUT })
-  }
-
   firstArticleCard(): Locator {
     return this.articleCards.first()
   }
@@ -48,35 +36,5 @@ export class TrendsPage {
     const articleCard = this.articleCards.filter({ hasText: title }).first()
     await expect(articleCard).toBeVisible({ timeout: TIMEOUT })
     await articleCard.click()
-  }
-
-  async expectArticleCount(count: number): Promise<void> {
-    await expect(this.articleCards).toHaveCount(count)
-  }
-
-  async expectQiitaIconCount(count: number): Promise<void> {
-    await expect(this.qiitaIcons).toHaveCount(count)
-  }
-
-  async expectZennIconCount(count: number): Promise<void> {
-    await expect(this.zennIcons).toHaveCount(count)
-  }
-
-  async expectHatenaIconCount(count: number): Promise<void> {
-    await expect(this.hatenaIcons).toHaveCount(count)
-  }
-
-  async waitForUrl(url: string | RegExp): Promise<void> {
-    await expect(this.page).toHaveURL(url, { timeout: TIMEOUT })
-  }
-
-  expectQueryParamNull(paramName: string): void {
-    const url = new URL(this.page.url())
-    expect(url.searchParams.get(paramName)).toBeNull()
-  }
-
-  expectQueryParamPresent(paramName: string): void {
-    const url = new URL(this.page.url())
-    expect(url.searchParams.get(paramName)).not.toBeNull()
   }
 }
