+++ application/packages/e2e/src/scenario/article-browse.test.ts
@@ -1,0 +1,58 @@
+import { expect, test } from '../fixtures'
+import * as articleHelper from '../helper/article'
+import { ArticleDrawer } from '../pom/components/article-drawer'
+import { SUPPORTED_ARTICLE_URL_PATTERN } from '../pom/constants'
+import { TrendsPage } from '../pom/trends-page'
+
+const ARTICLE_COUNT = 10
+
+test.describe('記事閲覧シナリオ', () => {
+  const createdArticleIds: bigint[] = []
+
+  test.beforeAll(async ({ rdb }) => {
+    const articles = await Promise.all(
+      Array.from({ length: ARTICLE_COUNT }, () => articleHelper.createArticle(rdb)),
+    )
+    createdArticleIds.push(...articles.map((a) => a.articleId))
+  })
+
+  test.afterAll(async ({ rdb }) => {
+    await articleHelper.cleanUp(rdb, createdArticleIds)
+  })
+
+  test.beforeEach(async ({ page }) => {
+    const trendsPage = new TrendsPage(page)
+    await trendsPage.goto()
+    await trendsPage.waitForArticleCards()
+  })
+
+  test('記事一覧から記事詳細を閲覧し、再び記事一覧に戻る', async ({ page }) => {
+    const trendsPage = new TrendsPage(page)
+    const articleCard = trendsPage.firstArticleCard()
+    await expect(articleCard).toBeVisible()
+
+    await trendsPage.openFirstArticle()
+
+    const drawer = new ArticleDrawer(page)
+    await drawer.waitOpen()
+    await drawer.close()
+    await drawer.expectClosed()
+
+    // 記事一覧に戻っていることを確認(記事カードが表示されていること)
+    await expect(articleCard).toBeVisible()
+  })
+
+  test('記事一覧から記事詳細を閲覧し、その実際の記事を閲覧する', async ({ page }) => {
+    const drawer = new ArticleDrawer(page)
+    await drawer.mockWindowOpen()
+
+    const trendsPage = new TrendsPage(page)
+    await trendsPage.openFirstArticle()
+    await drawer.waitOpen()
+    await drawer.clickReadArticle()
+    const openedUrl = await drawer.getLastOpenedUrl()
+
+    // 記事URLが開かれることを確認
+    expect(openedUrl).toMatch(SUPPORTED_ARTICLE_URL_PATTERN)
+  })
+})
